requirements.txt
streamlit
easyocr
Pillow
pandas
numpy
PyMuPDF
